Fix height overflow issue
Due to margin collapsing, the margin in #viewer-messages became
collapsed with the parent body, which caused the height overflow and
triggering the vertical scrollbar. This is alleviated by converting it
to padding instead.

Also, in Firefox 19, even after the padding fix, the height still
overflowed slightly. The fix is to hide the vertical scrollbar as it is
not needed here.

The converted document iframe is hidden until needed.